Fix deletion logic by foregoing data-index and using title instead.
This commit changes the deletion logic to forgo data-index as
filter/splice was moving the array elements around and the data-index quickly
became redundant. This would cause books that were supposed to be removed to appear again.
This is changed to delete books based on their title alone. This is independent of index
and therefore is more rigid.